Fix glyphs name.
