+++ generator/glyphs/name.py
@@ -1,0 +1,91 @@
+class NerdFonts:
+    ALL = [
+        # Font Awesome
+        (0xE216, "font-awesome_ext_slash"),
+        (0xF004, "font-awesome_heart"),
+        (0xF015, "font-awesome_home"),
+        (0xF046, "font-awesome_check"),
+        (0xF0A2, "font-awesome_bell"),
+        (0xF111, "font-awesome_circle"),
+        (0xF118, "font-awesome_smile"),
+        (0xF119, "font-awesome_frown"),
+        (0xF140, "font-awesome_bullseye"),
+        (0xF14E, "font-awesome_compass"),
+        (0xF182, "font-awesome_female"),
+        (0xF183, "font-awesome_male"),
+        (0xF185, "font-awesome_sun"),
+        (0xF221, "font-awesome_venus"),
+        # Octicons
+        (0xF420, "octicons_question"),
+        (0xF42E, "octicons_check"),
+        (0xF467, "octicons_x"),
+        (0xF46D, "octicons_home"),
+        (0xF475, "octicons_ellipsis"),
+        (0xF49A, "octicons_bell"),
+        # Material
+        (0xF52A, "material_alpha"),
+        (0xF564, "material_at"),
+        (0xF56D, "material_backspace"),
+        (0xF599, "material_bell"),
+        (0xF5A0, "material_alpha"),
+        (0xF5D3, "material_box"),
+        (0xF62B, "material_chart-histogram"),
+        (0xF68A, "material_compass"),
+        (0xF6BF, "material_delete"),
+        (0xF6C1, "material_delta"),
+        (0xF6C7, "material_diamond"),
+        (0xF6E6, "material_earth"),
+        (0xF6FB, "material_equal"),
+        (0xF7D0, "material_heart"),
+        (0xF7DB, "material_home"),
+        (0xF7E8, "material_image"),
+        (0xF873, "material_minus"),
+        (0xF8C8, "material_open-in-new"),
+        (0xF8EF, "material_percent"),
+        (0xF8FE, "material_pi"),
+        # Moved Material
+        (0xE814, "material_plus"),
+        (0xE89F, "material_sigma"),
+        (0xE918, "material_tie"),
+        (0xE935, "material_triangle"),
+        (0xE9DC, "material_bullseye"),
+        (0xE9E5, "material_copyright"),
+        (0xEA26, "material_lambda"),
+        (0xEA9E, "material_gradient"),
+        (0xEAC2, "material_asterisk"),
+        (0xEAE0, "material_hook"),
+        (0xEAE2, "material_infinity"),
+        (0xEB04, "material_prescription"),
+        (0xEB23, "material_tilde"),
+        (0xEB38, "material_cancel"),
+        (0xEB62, "material_square"),
+        (0xEB63, "material_circle"),
+        (0xEBE0, "material_null"),
+        (0xEBE3, "material_pipe"),
+        (0xEBE9, "material_ring"),
+    ]
+
+    NOT_USING = [
+        0x2600,
+        0x260E,
+        0x26AA,
+        0x26AB,
+        0x27A1,
+        0x2B05,
+        0x2B06,
+        0x2B07,
+    ]
+
+
+class NotoEmoji:
+    ALL = [
+        (0x1F3E0, "emoji_house"),
+        (0x1F3F0, "emoji_castle"),
+        (0x1F3F3, "emoji_black-flag"),
+        (0x1F3F4, "emoji_white-flag"),
+        (0x1F478, "emoji_princess"),
+        (0x1F5E1, "emoji_dagger"),
+        (0x1F934, "emoji_prince"),
+        (0x1FA91, "emoji_chair"),
+        (0x1FA9D, "emoji_hook"),
+    ]
