+++ generator/glyphs/ranges.py
@@ -1,0 +1,51 @@
+from configurator.glyphs.ranges import UnicodeRanges
+
+
+class IosevkaExtended:
+    ALL: UnicodeRanges = [
+        (0x0000, 0x052F),
+        (0x0E00, 0x0E7F),
+        (0x1AB0, 0x1AFF),
+        (0x1D00, 0x243F),
+        (0x2460, 0x2BFF),
+        (0x2C60, 0x2C7F),
+        (0x2DE0, 0x2E7F),
+        (0xA640, 0xA69F),
+        (0xA700, 0xA7FF),
+        (0xAB30, 0xAB6F),
+        (0xFB00, 0xFB4F),
+        (0xFFF0, 0xFFFF),
+        (0x1D400, 0x1D7FF),
+        (0x1F100, 0x1F1FF),
+        (0x1F300, 0x1F5FF),
+        (0x1F780, 0x1F8FF),
+        (0x1FB00, 0x1FBFF),
+    ]
+
+
+class BIZUDGothic:
+    HALF_WIDTH_KATAKANA: UnicodeRanges = [(0xFF61, 0xFFDF)]
+    FULL_WIDTH_HIRAGANA: UnicodeRanges = [(0x3000, 0x31FF)]
+    FULL_WIDTH_LATIN: UnicodeRanges = [(0xFF00, 0xFF60), (0xFFE0, 0xFFEF)]
+    FULL_SIZE_GLYPHS: UnicodeRanges = [
+        (0x3200, 0x4DBF),
+        (0x4E00, 0x9FFF),
+        (0xF900, 0xFAFF),
+        (0xFE30, 0xFE4F),
+        (0x20000, 0x2A6DF),
+        (0x2F800, 0x2FA1F),
+    ]
+
+
+class NerdFonts:
+    ALL: UnicodeRanges = [(0xE000, 0xFD46)]
+
+
+class NotoEmoji:
+    ALL: UnicodeRanges = [
+        (0x1F300, 0x1F64F),
+        (0x1F680, 0x1F6FF),
+        (0x1F780, 0x1F7FF),
+        (0x1F900, 0x1F9FF),
+        (0x1FA70, 0x1FAFF),
+    ]
